move widget related types to sdk
