+++ plugins/sdk/src/index.ts
@@ -3,6 +3,7 @@ export * from './lib/sdk.module';
 export * from './lib/components';
 export * from './lib/layout';
 export * from './lib/services';
+export * from './lib/widgets';
 
 export * from './lib/tokens';
 export * from './lib/types';
