+++ plugins/sdk/src/lib/widgets/add-widget-dialog/add-widget.dialog.html
