+++ plugins/sdk/src/lib/widgets/add-widget-dialog/index.ts
@@ -1,0 +1,1 @@
+export * from './add-widget.dialog';
