+++ plugins/sdk/src/lib/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from './add-widget-dialog';
+export * from './widget-container';
+export * from './widget';
