+++ plugins/sdk/src/lib/widgets/widget-container/index.ts
@@ -1,0 +1,1 @@
+export * from './widget-container.component';
