+++ src/app/pages/dashboard/dashboard.component.html
@@ -1,6 +1,6 @@
 <h1>Dashboard</h1>
 <div>
-  <button mat-button (click)="onAddWidget()">Add</button>
+  <button mat-stroked-button (click)="onAddWidget()">Configure</button>
 </div>
 
 <mat-grid-list cols="4" rowHeight="200px">
